requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
jedi==0.19.1
jieba3k==0.35.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.3
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
snowflake-id==0.0.5
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
mattsollamatools==0.0.16
flask==3.0.2
pymysql==1.1.0
